# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TokenPairDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Authentication"]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TokenPairDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Authentication"]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TokenPairDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Authentication"]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenDto" } }
          }
        },
        "responses": {
          "204": { "description": "" },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Authentication"]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthUserDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Authentication"]
      }
    },
    "/api/v1/me/profile": {
      "get": {
        "operationId": "ProfileController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Profile"]
      },
      "patch": {
        "operationId": "ProfileController_update_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProfileDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Profile"]
      }
    },
...
  },
  "info": {
    "title": "PLANDA Product API",
    "description": "Shared API for web, iOS, Android, developer, broker and admin clients",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Validation failed" },
          "details": { "type": "array", "items": { "type": "string" } },
          "requestId": { "type": "string", "example": "4f12bc5f-1a5e-4d14-bf14-6c7bb19ea413" }
        },
        "required": ["statusCode", "code", "message", "requestId"]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "buyer@example.test" },
          "password": { "type": "string", "example": "DevelopmentOnly!123", "minLength": 12 }
        },
        "required": ["email", "password"]
      },
      "OrganizationMembershipDto": {
        "type": "object",
        "properties": {
          "organizationId": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["OWNER", "ADMIN", "MEMBER"] }
        },
        "required": ["organizationId", "role"]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["BUYER", "DEVELOPER_MEMBER", "BROKER", "ADMIN", "SUPER_ADMIN"]
            }
          },
          "organizationMemberships": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrganizationMembershipDto" }
          }
        },
        "required": ["id", "email", "roles", "organizationMemberships"]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "accessTokenExpiresIn": { "type": "number", "example": 900 },
          "refreshTokenExpiresIn": { "type": "number", "example": 2592000 },
          "user": { "$ref": "#/components/schemas/AuthUserDto" }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresIn",
          "refreshTokenExpiresIn",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "buyer@example.test" },
          "password": { "type": "string", "example": "DevelopmentOnly!123" }
        },
        "required": ["email", "password"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Opaque-to-client rotating refresh JWT"
          }
        },
        "required": ["refreshToken"]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email", "readOnly": true },
          "fullName": { "type": "object", "nullable": true },
          "phone": { "type": "object", "nullable": true },
          "preferredLanguage": {
            "type": "string",
            "enum": ["TR", "EN", "AR", "RU"],
            "nullable": true
          },
          "preferredCurrency": { "type": "object", "nullable": true },
          "createdAt": { "format": "date-time", "type": "string", "readOnly": true },
          "roles": {
            "type": "array",
            "readOnly": true,
            "items": {
              "type": "string",
              "enum": ["BUYER", "DEVELOPER_MEMBER", "BROKER", "ADMIN", "SUPER_ADMIN"]
            }
          }
        },
        "required": ["id", "email", "createdAt", "roles"]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": { "type": "object", "nullable": true, "minLength": 2, "maxLength": 160 },
          "phone": { "type": "object", "nullable": true, "example": "+905551112233" },
          "preferredLanguage": {
            "type": "string",
            "enum": ["TR", "EN", "AR", "RU"],
            "nullable": true
          },
          "preferredCurrency": { "type": "object", "nullable": true, "example": "TRY" }
        }
      },
...
    }
  }
}